Allow specifying an init function for :redis_provider
Since we cannot be sure in which order a release will start it's
dependencies we can make sure any dependencies required to use the
`:redis_provider` (like :poolboy) are started before toniq.
This is done by adding a `:redis_init = {Application, :ensure_all_started, [:poolboy]}`
to the config.

This is a hackish way of doing things but Erlang releases expect each
otp app to statically define it's dependencies. This means if `toniq`
requires `poolboy`, `toniq` MUST define the dependency otherwise there's
no guarantee that `poolboy` will be started first.

Since Toniq starts working immediately without waiting for the external
redis client to be injected it will fail when started as a release.
This apatch allows us to await for some function to return before
starting `toniq`; meaning `poolboy` will be allowed to startup
before we continue.